.spec.jsx files have been edited, components are arranged in folders. Added handleSubmit and componentDidUpdate methods for Regform
