+++ src/components/registration-form/RegForm.jsx
@@ -1,17 +1,27 @@
 import React from 'react';
-import Button from './Button';
+import Button from '../button/Button';
 import PropTypes from "prop-types";
+import { AuthContext } from '../../context/AuthContext';
 
 class RegForm extends React.Component {
 	static propTypes = {
 		navigateTo: PropTypes.func
 	}
 
+	static contextType = AuthContext;
+
 	state = {
 		email: ``,
 		name: ``,
 		password: ``
 	};
+
+	handleSubmit = e => {
+		e.preventDefault();
+		const { email, password } = this.state;
+		this.context.logIn(email, password);
+	};
+
 
 	handleChange = e => {
 		this.setState({ [e.target.name]: e.target.value });
@@ -25,7 +35,7 @@ class RegForm extends React.Component {
 			<>
 				<div className="form">
 					<div className="form__title">Регистрация</div>
-					<form name='RegForm' onSubmit={() => navigateTo("map")}>
+					<form name='RegForm' onSubmit={this.handleSubmit}>
 						<div className="form__item">
 							<label>
 								<span>Email*</span>
@@ -55,6 +65,12 @@ class RegForm extends React.Component {
 				</div>
 			</>
 		)
+	}
+
+	componentDidUpdate() {
+		if (this.context.isLoggedIn) {
+			this.props.navigateTo('map')
+		}
 	}
 }
