+++ src/pages/login/Login.js
@@ -1,24 +1,24 @@
 import React from 'react';
-import RegForm from './../components/RegForm';
+import LoginForm from '../../components/login-form/LoginForm';
 import PropTypes from "prop-types";
 
-function Reg(props) {
+function Login(props) {
 	const { navigateTo } = props;
 
 	return (
 		<>
 			<main className='maincontent'>
 				<div className='container'>
-					<h1>Регистрация</h1>
-					<RegForm navigateTo={navigateTo} />
+					<h1>Логин</h1>
+					<LoginForm navigateTo={navigateTo} />
 				</div>
 			</main>
 		</>
 	)
 }
 
-Reg.propTypes = {
+Login.propTypes = {
 	navigateTo: PropTypes.func
 }
 
-export default Reg;
+export default Login;
